% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cran_sample}
\alias{cran_sample}
\title{Sample data-indexed dataset.}
\format{A data frame with 1462 observations of 3 variables:
\describe{
\item{date}{\code{Date}, date of the observation}
\item{package}{\code{character}, name of the package}
\item{count}{\code{integer}, number of downloads according to the RStudio logs}
}}
\source{
Downloaded using the
\href{https://CRAN.R-project.org/package=cranlogs}{cranlogs} package.
}
\usage{
cran_sample
}
\description{
Daily downloads from RStudio CRAN mirror for \strong{ggplot2} and \strong{dplyr} for
the years 2015 and 2016.
}
\keyword{datasets}
