% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_weekend}
\alias{data_weekend}
\title{Sample data-indexed dataset.}
\format{A data frame with 731 observations of 5 variables:
\describe{
\item{date}{\code{Date}, date of the observation}
\item{observation}{\code{character}, indicates if this is a "weekend" or "weekday"}
\item{probability}{\code{double}, "model" probability that this is a "weekend"}
\item{prediction}{\code{character}, "model" prediction of type of day}
\item{strength}{\code{double}, "model" probability that this is a "weekend"}
}}
\usage{
data_weekend
}
\description{
A dataset to use to test data-based visualizations.
}
\keyword{datasets}
