% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose_date}
\alias{decompose_date}
\alias{mutate_ggocto}
\title{Decompose a date into components}
\usage{
decompose_date(date, use_iso = FALSE, use_year = TRUE)

mutate_ggocto(.data, date = .data$date, use_iso = FALSE, use_year = TRUE)
}
\arguments{
\item{date}{\code{Date}, date to decompose}

\item{use_iso}{\code{logical}, use ISO weeks (vs. US-style weeks)}

\item{use_year}{\code{logical}, use \code{year} as a variable}
}
\value{
\link{decompose_date}:
\code{tbl_df} with variables:
\describe{
\item{.year}{\code{integer}, year to which the week belongs}
\item{.week}{\code{integer}, number of week}
\item{.day_of_week}{\code{integer}, identifies the day in the week}
}

\code{\link[=mutate_ggocto]{mutate_ggocto()}}:
\code{.data}, with additional columns from \code{\link[=decompose_date]{decompose_date()}}
}
\description{
Decompose a date into components
}
